+++ docs/python/api-reference/mcp_use_server_auth_bearer.mdx
@@ -27,7 +27,8 @@ Base class for bearer token authentication.
 
     Example:
         ```python
-        from mcp_use.server import MCPServer, BearerAuthProvider, AccessToken
+        from mcp_use.server import MCPServer
+        from mcp_use.server.auth import BearerAuthProvider, AccessToken
 
         class MyAuthProvider(BearerAuthProvider):
             async def verify_token(self, token: str) -> AccessToken | None:
