+++ docs/python/api-reference/mcp_use_server_auth_models.mdx
@@ -29,7 +29,7 @@ Validated access token with user claims.
 
     Example:
         ```python
-        from mcp_use.server import get_access_token
+        from mcp_use.server.auth import get_access_token
 
         @server.tool()
         def whoami() -> str:
