+++ libraries/python/mcp_use/server/__init__.py
@@ -1,4 +1,3 @@
-from .auth import AccessToken, AuthenticationError, AuthMiddleware, BearerAuthProvider, get_access_token, require_auth
 from .context import Context
 from .middleware import Middleware, TelemetryMiddleware
 from .router import MCPRouter
@@ -14,11 +13,4 @@ __all__ = [
     "Context",
     "Middleware",
     "TelemetryMiddleware",
-    # Auth exports
-    "BearerAuthProvider",
-    "AccessToken",
-    "AuthMiddleware",
-    "AuthenticationError",
-    "get_access_token",
-    "require_auth",
 ]
